+++ plugins/cost-insights/src/components/ProjectGrowthAlertCard/ProjectGrowthAlertCard.test.tsx
@@ -27,6 +27,9 @@ import { AlertCost } from '../../types';
 import { defaultCurrencies } from '../../utils/currency';
 import { findAlways } from '../../utils/assert';
 
+// suppress recharts componentDidUpdate deprecation warnings
+jest.spyOn(console, 'warn').mockImplementation(() => {});
+
 const engineers = findAlways(defaultCurrencies, c => c.kind === null);
 
 const MockProject = 'test-project-1';