+++ plugins/kubernetes-backend/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @backstage/plugin-kubernetes-backend
 
+## 0.2.2
+
+### Patch Changes
+
+- Updated dependencies [38e24db00]
+- Updated dependencies [e3bd9fc2f]
+- Updated dependencies [12bbd748c]
+- Updated dependencies [83b6e0c1f]
+- Updated dependencies [e3bd9fc2f]
+  - @backstage/backend-common@0.4.0
+  - @backstage/config@0.1.2
+  - @backstage/catalog-model@0.5.0
+
 ## 0.2.1
 
 ### Patch Changes